chore: move variable scope up
